p4i integration Integrate 204135 from 4.1 to main: Fixes:    synchronization issues Task:     98081, 91847 RevBy:    paul Details:  Synchronize decoration painting with widget painting to 	  remove flickering. 	  Grab display lock when blitting to screen. 	  Separate lock for communication and backingstore operations 	  to avoid deadlock. 	  Avoid using shared memory and locking on server side client. 	  Don't do unecessary sleeps when connecting to server pipe 	  (start-up synchronization).
     		  All this and more in a single submit for your convenience.

[git-p4: depot-paths = "//depot/qt/main/": change = 204136]

+++ src/gui/embedded/qwindowsystem_p.h
@@ -89,7 +89,7 @@ private:
     void set_altitude(const QWSChangeAltitudeCommand *);
     void set_opacity(const QWSSetOpacityCommand *);
     void request_focus(const QWSRequestFocusCommand *);
-    void request_region(int winId, int shmid, int windowtype, QRegion, QWSWindow* =0);
+    void request_region(int winId, quint64 shmid, int windowtype, QRegion, QWSWindow* =0);
     void repaint_region(int winId, bool opaque, QRegion);
     void destroy_region(const QWSRegionDestroyCommand *);
     void name_region(const QWSRegionNameCommand *);
